Add files via upload
Input

The first line contains a word s — it consists of uppercase and lowercase Latin letters and possesses a length from 1 to 100.
Output

Print the corrected word s. If the given word s has strictly more uppercase letters, make the word written in the uppercase register, otherwise - in the lowercase one.